==> pandas/1_Pandas I.ipynb <==
import numpy as np
import pandas as pd

1. ¿Qué diferencia hay entre una Serie y un DataFrame? una es unidimensional y otra bidimensional
2. Momento de trabajar con Series. Contestad a las siguientes preguntas:
Crea dos listas:
Una que se llame "valores" que tenga 7 números aleatorios.
Otra que se llame "indices" que tenga 7 strings que se llamen como los días de la semana


valores1 = np.random.randint (0,10, 7)
valores1

valores= pd.Series (valores1)
valores

indices1= ['lunes', 'martes', 'miercoles', 'jueves', 'viernes', 'sabado', 'domingo']
indices= pd.Series (indices1)
indices

A partir de estas dos listas cread una Serie donde la variable "valores" sean los valores de la Serie y la variable indices sean los índices de la Serie.

serie_1= pd.Series (valores1, indices)
serie_1

A partir de las dos listas anteriores cread un diccionario usando un bucle for. Para hacer esto tenemos el método zip que nos permite iterar por dos listas a la vez. Más informción sobre este método aquí.

diccionario = dict(zip(indices, valores1))
diccionario

diccionario_s=  pd.Series(diccionario)
diccionario_s

diccionario_s.shape

diccionario_s.dtype

diccionario_s.size

diccionario_s.index

diccionario_s.values

diccionario_s['lunes']

diccionario_s['lunes': 'miercoles']

diccionario_s[['lunes','domingo']]

3. Es el turno de los DataFrames
    - Cread un array de dos dimensiones con 4 filas y 3 columnas.

array3= np.random.randint (10,50, (4,3))
array3

paises= pd.DataFrame (array3, columns= ['España', 'Francia', 'Italia'])
paises

df = pd.read_csv("medallas.csv", index_col=["País"])
df

- Usando el loc extraed todas las columnas de la fila.
- Usando el loc extraed todas las columnas de China e Irán.
- Usando el loc extraed solo las columnas de "Oro", "Plata" y "Bronce" de China e Irán.


df.loc [['Tailandia'], :]

df.loc [['China', 'Irán'], :]

df.loc [['China', 'Irán'], ['Oro', 'Plata', 'Bronce']]

- Usando el iloc extraed la información de la India. Devolved todas las columnas.
- Usando el iloc extraed todas las columnas de los paises que estén en posición impar. La primera que nos tendría que salir es Japón y la última Corea del Norte.
- Utilizando el código del ejercicio anterior, seleccionad unicamente las columnas "Oro", "Plata" y "Bronce".

df = pd.read_csv("medallas.csv", index_col=["País"])
df


df.iloc[[5],]


df.iloc[1: :2, :]

df.iloc[1: :2, 1:]

Contestad a las siguientes preguntas:
- ¿Qué país consiguió más de 200 medallas de oro?
- ¿Qué país consiguió más de 200 medallas de oro y más de 1000 de plata?
- ¿Qué país consiguió más de 200 medallas de oro o más de 700 de bronce?

df.loc [df.Oro>200]

df.loc [(df.Oro>200) & (df.Plata>1000)]

df.loc [(df.Oro>200) | (df.Bronce>70)]

- Como vemos hay una columna que se llama Unnamed: 0 que no nos interesa. Eliminad esa columna.


df.drop(['Unnamed: 0'], axis=1)

- Cread una nueva columma que se llame "total" que sea el resultado de la suma de todas las medallas obtenidas por cada país.

df['suma_medallas']=(df['Oro']+df['Plata']+df['Bronce'])
df.head()
