==> pandas/2_pandas-IIM.ipynb <==
import pandas as pd
import numpy as np
import pickle


# Id cargando todos los ficheros en vuestro jupyter.
# Si os da error leer el pickle será por la versión de pandas. Podéis guardar el csv como pickle y 
# cargarlo de ahí.

# fichero quejas.csv
df_csv = pd.read_csv("quejas.csv")
df_csv.head(2)


# fichero quejas.csv
# cómo nos daba error al ejectutar "UnicodeDecodeError: 'utf-8' codec can't decode byte 0xf3 in position 170:
# invalid continuation byte", buscando en internet, hemos intentado con 'encoding:latin-1', 'encoding:ISO-8859-1' 
# y por posibles errores en la lectura de caracteres 'sep = ;'. (1º intento en cell siguiente)
# de 
df_csv = pd.read_csv("quejas.csv", sep = ";", encoding = "ISO-8859-1")
df_csv.head(2)

# fichero quejas.json
df_json = pd.read_json('quejas.json')
df_json.head(4)

# fichero quejas2.xlsx
df_xlsx = pd.read_excel('quejas2.xlsx')
df_xlsx.head(3)

Si os da error leer el pickle será por la versión de pandas. Podéis guardar el csv como pickle y cargarlo de ahí.

# fichero algas.pickle
df_csv_algas = pd.read_pickle("algas.pickle")
df_csv_algas.head(3)

df_xlsx.to_pickle('df_xls.pickle')

pd.read_pickle('df_xls.pickle')

# Para poner el práctica el read_clipboard elegir alguna tabla que os podaís encontrar en google y 
# usad dicho método para traer los datos al jupyter notebook. Una vez que los hayáis cargado, guardad 
# esos datos en los siguientes formatos:
        # csv
        # excel
        # json
        # pickle

#https://www.ibm.com/docs/es/tap/3.5.0?topic=tables-example-fact-table-associated-dimensions

df_clipboard.head()

#sale con la direccion web pero no con la tabla

SPEC_ID
Número de nivel
Nivel 1 SPEC_ID
Nivel 2 SPEC_ID
Nivel 3 SPEC_ID
Nivel 4 SPEC_ID
Mundo	1	Mundo	N/D	N/D	N/D
América del Norte	2	Mundo	América del Norte	N/D	N/D
EMEA	2	Mundo	EMEA	N/D	N/D
APAC	2	Mundo	APAC	N/D	N/D
Estados Unidos	3	Mundo	América del Norte	Estados Unidos	N/D
Canadá	3	Mundo	América del Norte	Canadá	N/D
Nevada	4	Mundo	América del Norte	Estados Unidos	Nevada
Texas	4	Mundo	América del Norte	Estados Unidos	Texas

df_clipboard = pd.read_clipboard()
