==> pandas/3_panda III_métodos.ipynb <==
import pandas as pd


ataques= pd.read_csv ("attacks.csv")


¿Cuántas filas y columnas tiene el DataFrame?
Mostrad las primeras cinco filas y las últimas cinco filas del DataFrame.
Mostrad 2 filas al azar del DataFrame.
Mostrad 2 filas al azar del DataFrame.

ataques.shape

ataques.head()


ataques.tail()

ataques.sample(2)

ataques.columns

¿Cuáles son los principales estádisticos del DataFrame? Mostrad por separado las variables categóricas y las variables numéricas.

ataques.describe()

ataques.describe (include= 'object')

Cuál es el porcentaje de valores nulos en cada columna?


porcentaje_nulos=(ataques.isnull().sum()/ataques.shape[0])*100
porcentaje_nulos

Sacad los valores únicos y su frecuencia de las variables numéricas. Almacenad los resultados en un diccionario. Uno para cada variable.

ataques.dtypes


ataques['Unnamed: 0'].unique()

ataques['Year'].unique()

ataques['original order'].unique()

frecuencia=ataques['Unnamed: 0'].value_counts(), ataques['Year'].value_counts(), ataques['original order'].value_counts()
frecuencia


diccionario_unnamed = dict(ataques['Unnamed: 0'].value_counts())
diccionario_unnamed


diccionario_year=dict(ataques['Year'].value_counts())
diccionario_year


diccionario_original_order= dict(ataques['original order'].value_counts())
diccionario_original_order

Cambiad el nombre de las columnas para que todas estén en minúsculas y reemplazad los espacios por _.

columnas_nuevas = {col :  col.lower() for col in ataques.columns}
columnas_nuevas

ataques.rename (columns = columnas_nuevas, inplace = True)
ataques.head(1)

columnas_nuevas2 = {col :  col.replace(' ', '_') for col in ataques.columns}

ataques.rename (columns = columnas_nuevas2, inplace = True)
ataques.head(5)

Eliminad las columnas de Unnamed: 22 y Unnamed: 23, Case Number.1, Case Number.2, original order, Investigator or Source, pdf, href formula.
Establece la columna Case Number como índice del DataFrame.

ataques.drop (['unnamed:_22','unnamed:_23', 'case_number.1', 'original_order', 'investigator_or_source','pdf', 'href_formula'], axis=1, inplace=True)
ataques.head(5)

Establece la columna Case Number como índice del DataFrame.

ataques.set_index ('case_number')

ataques.to_csv('ataques.csv')
