==> pandas/4_Pandas_IV_filtrado_M.ipynb <==
import pandas as pd
import numpy as np

df=pd.read_csv('ataques.csv')

df.head()
df.shape

df_limpio= df.dropna()

¿Cuántas filas se han eliminado con esta línea de código? 

df_limpio.shape 
#se han eliminado 4612 filas

¿Cuál es el número de ataques que se han producido en Estados Unidos?

df_limpio.columns
#comprobamos como se llama la col que buscamos y es country

df_limpio.head()
df_limpio['country'].unique()
#comprobamos cómo esta escrito el pais que buscamos y es USA

#hacemos un filtro para q nos muestre solo lo de USA y con un shape vemos cuantos registros/casos hay
ataques_usa = df_limpio[df_limpio['country']=='USA']
ataques_usa.shape
#se han producido 719 ataques en USA

Y en España?
3.1 ¿En qué regiones se han producido los ataques (columna location) ?
3.2 ¿Cuántos ataques se han producido en cada área (columna area?

ataques_spain = df_limpio[df_limpio['country']=='SPAIN']
ataques_spain.shape
#se han producido 5 ataques en España

ataques_spain['location'].unique()
#sobre la tabla de spain filtramos por la localizacion para ver las regiones donde hubo ataques

ataques_spain.head(1)

#contamos los valores de cada area de españa para ver cuantos ataques se han producido en casa area
ataques_spain['area'].value_counts()

4. Extraed un DataFrame nuevo con los ataques que se han producido en Australia y Estados Unidos. 📌 Nota Este ejercicio se puede resolver usando dos condiciones de filtrado o usando el método isin(). Hacedlo con ambas herramientas.



df_paises= df_limpio[df_limpio['country'].isin (['AUSTRALIA','USA'])]
df_paises.head(1)

filtro1 = df_limpio[(df_limpio['country']=='USA') | (df_limpio['country']=='AUSTRALIA')]
filtro1.head(1)

4.1 ¿Cuántos ataques se han producido en cada país?



df_usa= df_limpio[df_limpio['country'].isin (['USA'])]
df_usa.shape
#se han producido 719 ataques

#lo comprobamos tambien usando los values_count
casos_usa= df_usa['case_number'].value_counts() 
casos_usa
#en usa ha habido 719 ataques

ataques_australia = df_limpio[df_limpio['country']=='AUSTRALIA']
ataques_australia.shape

#en Australia ha habido 324 ataques

#hecho con value counts
casos_australia= ataques_australia['country'].value_counts() 
casos_australia

4.2 ¿Ha habido ataques en Australia a gente que estuviera remando (lo encontrareis como Paddling en el DataFrame en la columna activity)?


filtro_paddle = ataques_australia[ataques_australia['activity']=='Paddling']
filtro_paddle.head()
#no ha habido ataques haciendo Paddling

#lo comprobamos por otro camino y el resultado es el mismo
df_limpio[(df_limpio['country']=='AUSTRALIA') & (df_limpio['activity']=='Paddling')]

4.3 ¿Cuántos ataques se han producido en Australia antes del 2000?

4.4 ¿Y después del 2000?

antes_2000= ataques_australia[ataques_australia['year']<2000]
antes_2000.shape
#se han producido en Australia 151 ataques antes del año 2000

despues_2000= ataques_australia[ataques_australia['year']>=2000]
despues_2000.shape
#se han producido en Australia 173 después del año 2000

5. Extraed unicamente los ataques a mujeres.

ataques_mujeres=df_limpio [df_limpio['sex_']=='F']
ataques_mujeres.head(2)

6. ¿Cuántas mujeres fueron atacadas en España?

#1 mujer atacada en España
atacadas_spain= ataques_mujeres[ataques_mujeres['country']=='SPAIN']
atacadas_spain.shape

7. Extraed los ataques que se hayan producido en el año 2000, 1997 y 2006. Utilizad el método isin() para resolver este ejercicio.

df_años= df_limpio[df_limpio['year'].isin ([1997,2000,2006])]
df_años.shape
#92 ataques

8. Si extraemos los valores únicos de la columna species nos podemos dar cuenta que está un poco sucia y extraer informacion útil de la misma se puede convertir en algo difícil. En este ejercicio tendréis que extraer los ataques que han ocurrido por tiburón blanco (lo encontraremos en la columna species como algo que contenga white o White). Usad el método str.contains() para resolver este ejercicio.

patron = '(white|White)'

df_blanco= df_limpio[df_limpio['species_'].str.contains( patron , regex=True)]
df_blanco.head(2)
df_blanco.shape
#379 ataques de tiburon blanco



9. Extraed ahora los ataques por tiburón blanco o tigre. Utilizad el método str.contains() para resolver este ejercicio.

#9. Extraed ahora los ataques por tiburón blanco o tigre. Utilizad el método str.contains() para resolver este ejercicio.

patron2 = '(tiger|Tiger)'

df_tigre_blanco= df[(df['species_'].str.contains (patron, regex=True)) & (df['species_'].str.contains (patron2, regex=True))]
df_tigre_blanco.shape
#2 ataques



10. Utilizando la columna de year cread una nueva columna llamada siglo que tiene siglo-XX o siglo-XXI. Donde el año sea menor que 2000 pondremos siglo-XX y el resto de los años siglo-XXI.




df_limpio["siglo"] = np.where(df_limpio["year"]< 2000,"SigloXX","SigloXXI")  
df_limpio.sample(5)

11. Guardad como un archivo csv nuevo. Lo usaremos en el pair programming de mañana.

df_limpio.to_csv('ataques_limpio.csv')
