==> pandas/5_pandas_V_union_datos_M.ipynb <==
Trabajaremos con tres ficheros diferentes:
- El csv que creamos en el ejercicio de pair programming Pandas IV
- El csv llamado attacks_historico donde tenemos información sobre los ataques de tiburón previos al año 1800, 
- El csv llamado attacks_adicional donde tenemos información sobre los ataques de tiburón. de nuevo,
El principal objetivo de este ejercicio de pair programming es acabar teniendo un único DataFrame con toda la información contenida en los tres csv anteriores.
Para llevar a cabo nuestro objetivo tendréis que seguir los siguientes pasos:

import pandas as pd
import numpy as np

1. Cargad los tres ficheros nombrados anteriormente y los exploramos para familizarizarnos con ellos y saber que información tienen en común y debatir con tu compañera como podríamos juntar toda la información.

df_limpio= pd.read_csv('ataques_limpio.csv')
df_adicional= pd.read_csv('attacks_adicional.csv')
df_historico= pd.read_csv('attacks_historico.csv')

2. Chequead que los nombres de las columnas ver si se llaman iguales entre dataframes . ¿Están todas en minúsculas? ¿Tienen espacios? En caso de que no sean todas iguales unificad el nombre de las columnas

df_historico.shape

df_historico.columns

df_limpio.columns


df_adicional.columns

#unificamos nombre col
col_nuevas_df_historico= {col :  col.lower().replace(' ', '_') for col in df_historico.columns}
col_nuevas_df_adicional= {col :  col.lower().replace(' ', '_') for col in df_adicional.columns}

df_historico.rename(columns=col_nuevas_df_historico, inplace=True)

df_adicional.rename(columns=col_nuevas_df_adicional, inplace=True)

#comprobamos
df_historico.columns

df_adicional.columns

3. En este momento nos habremos dado cuenta que el DataFrame que generamos en el pair programming Pandas IV es similar al DataFrame de los datos historicos. Si recordamos, eliminamos algunas columnas en la sesión III que no eran útiles, pero esas columnas están todavía en el histórico. Si no fuera por eso podríamos unir esos dataframes. Realizad los cambios que creais necesarios para que los dataframes sean iguales y luego juntadlos usando el mejor método.

df_historico.drop (['unnamed:_22','unnamed:_23', 'case_number.1', 'original_order', 'investigator_or_source','pdf', 'href_formula'], axis=1, inplace=True)

df_historico.shape

df_limpio.shape
#vemos que hay tres col de diferencia por lo que vamos a compararlas para ver si alguna es prescindible

df_limpio['Unnamed: 0'].unique

df_limpio['Unnamed: 0.1'].unique

df_limpio['unnamed:_0'].unique

df_limpio['Unnamed: 0.1'][1497]

df_limpio['unnamed:_0'][1497]

df_limpio['Unnamed: 0'][1497]

#conclusion: de las tres columnas que se llaman parecido solo dos son identicas asi que solo eliminamos una, eliminamos unnamed:0
df_limpio.drop (['Unnamed: 0'], axis=1, inplace=True)

#además por si acaso eliminamos duplicadas 
df_limpio.drop_duplicates(inplace=True)
df_limpio.head(1)

df_historico.head(1)

df_limpio.head(1)

#los unimos con un concat
df_concat = pd.concat([df_limpio,df_historico], axis=0, join='outer',ignore_index=True)
df_concat.head(1)

#comprobamos dimensiones
print('- dimensiones de concat:', df_concat.shape,'-dimensiones historico:', df_historico.shape, '-dimensiones de limpio:', df_limpio.shape)

3. Es el momento de unir la información de todos los ataques de tiburón que acabamos de crear con el DataFrame de información adicional. Debatid que método es el más correcto para unir toda esta información y ejecutadlo.
📌 NOTA Cuando hagáis la unión considerad que la información que más nos interesa es la del DataFrame que creamos en el apartado anterior, lo cual determinará el tipo de unión.

df_adicional.head(1)


mergeado_izq = pd.merge(left=df_concat, right=df_adicional, how='left', on='case_number')

mergeado_izq.head(1)



#comprobamos dimensiones
print('- dimensiones de mergeado:', mergeado_izq.shape,'- dimensiones concat:', df_concat.shape)

5. Guardad el dataframe nuevo en un csv que usaremos mañana en la clase de pair programming

mergeado_izq.to_csv ('DF_resultado_PandasV.csv')
