==> pandas/6_pandasVI_groupby.ipynb <==
import pandas as pd

df = pd.read_csv ('attacks_completo .csv')

1. ¿Cuántos ataques hubo por país? Devuelve un DataFrame.
📌 NOTA Ordena los resultados de mayor a menor. Esto todavía no lo hemos visto, pero un DataFrame lo podemos ordenar con .sort_values(). Aquí tenéis una explicación de cómo hacerlo.


df_paises = df.groupby (['country'])['case_number'].count()
df_paises.head(10)

df_paises.sort_values(ascending=False, inplace=True)

df_paises= pd.DataFrame(df_paises).reset_index()

2. ¿Cuántos ataques ha habido por año? Devuelve un DataFrame. De nuevo ordena los resultados de mayor a menor.


df_year= df.groupby(['year'])['case_number'].count()
df_year.sort_values(ascending=False, inplace=True)

df_year= pd.DataFrame(df_year).reset_index()

3. ¿Cuántos ataques hubo por año y sexo? Devolved un DataFrame solo del año 2000 en adelante.


df_sexo_year= df.groupby (['sex_','year'])['case_number'].count()
df_sexo_year.sort_values (ascending=True, inplace=True)
df_sexo_year=pd.DataFrame(df_sexo_year).reset_index()
df_sexo_year

df_sexo_year2000= df_sexo_year[df_sexo_year['year']>2000]
df_sexo_year2000.head(2)

4. Primero extraed solo las filas que correspondan a USA, AUSTRALIA y SOUTH AFRICA. ¿Cuántos de los ataques fueron letales por sexo? Devuelve un DataFrame.


df_usa= df[df['country'].isin (['USA', 'AUSTRALIA', 'SOUTH AFRICA'])]
df_usa.head(1)

df_usa.columns

df_usa2= df_usa.groupby (['sex_','country', 'fatal_(y/n)']) ['case_number'].count()
df_usa_letal= pd.DataFrame(df_usa2)

df2= df_usa_letal.reset_index()
df2.head(5)



filtro= df2[df2['fatal_(y/n)']=='Y']
filtro


5. Usando el DataFrame creado en el ejercicio anterior. ¿En qué país hubo más ataques?


df4 = df2.groupby('country')['case_number'].sum()
df4.reset_index()

6. ¿Qué y cuántos tipos de ataques hubo por país?

df_tipos=df.groupby ('country')['type'].value_counts()

cuadro_tipos= pd.DataFrame (df_tipos)
cuadro_tipos.reset_index()
