==> pandas/7_Pandas_VII-appy(Mrt).ipynb <==
import numpy as np
import pandas as pd
import datetime
import re


df = pd.read_csv ('attacks_completo .csv')
df.sample(4)

1. Cread una columna nueva y una función que nos de el mes en el que ocurrió el ataque. Tened en cuenta que no todas las filas tienen la misma estructura y que puede que no haya la información de mes. En ese caso devolved un nulo (NaN).
💡 Pista 💡Podéis usar regex.

df['date'].values

#nos aseguramos de que tipo son los valores de la columna para ver si se puede hacer con datetime
x= df['date'][0]
x
type(x)

#como son strings la solución es usar regex y definimos el patron. Para mayor seguridad hacemos un patron
#en el que nos capture todo lo que está entre guiones
patron= '-\D{3}-'

#definimos la función y la comprobamos para luego aplicar a toda la columna
def dime_mes (col):
    try:
        x= re.findall ('-\D{3}-', col)
        for i in x:
            y= i.replace ('-', '')
        return y
    except:
        return np.nan

#y la comprobamos 
dime_mes (df['date'][1])

dime_mes ('1900-1905')

#aplicamos y creamos nueva columna
df['mes'] =df['date'].apply(dime_mes)

df.tail(1)

df.head(1)

df['mes'].unique()

2. Vamos con la columna fatal_(y/n). Lo primero que tenéis que hacer es evaluar los valores únicos que hay. Esperaríamos tener solo dos valores: y y n, pero tristemente no será así 😭. Cread una columna nueva y una función que devuelva únicamente y o n y que devuelva un nulo (NaN) si no se encuentra el valor.

df['fatal_(y/n)'].unique()

#definimos y comprobamos la funcion
def si_no (col):
    if col=='Y'or col=='y':
        return 'y'
    elif col== 'N'or col=='n':
        return 'n'
    else:
        return np.nan

si_no (df['fatal_(y/n)'][1])

si_no (df['fatal_(y/n)'][205])

#creamos nueva columnas
df['fatal2'] = df['fatal_(y/n)'].apply (si_no)

df.head(1)

3. Seguid la misma lógica que en el ejercicio anterior pero con la columna sex. Pero en este caso vamos a incluir un retito más: En este caso querremos tener dos valores únicos F y M. De nuevo en la columna hay otros carácteres raros como 'M ', 'lli', 'N', '.'. En esos casos, queremos que se reemplacen por M o F de forma aleatoria (tendréis que usar la librería random). En caso de que sea nulo, que la función devuelva nulo (np.nan).

df['sex_'].unique()

def sexo (col):
    if col=='F' or col=='f':
        return 'F'
    elif col=='M' or col=='m':
        return 'M'
    elif col== 'nan':
        return np.nan
    

df['sexo'] =df['sex_'].apply(sexo)

4. Una vez que hayáis terminado eliminad las columnas originales conservando las nuevas.


df_nuevo_ataques= df.drop (columns= ['fatal_(y/n)', 'sex_', 'date'] )

5. Guardad el dataframe en un csv que usaremos más adelante

df_nuevo_ataques.to_csv('nuevo_ataques.csv')
